+++ admin_ui/views.py
@@ -4,25 +4,22 @@ from django.conf import settings
 from django.contrib import messages
 from django.contrib.contenttypes.models import ContentType
 from django.contrib.auth.decorators import login_required, user_passes_test
-from django.db import models
-from django.db.models import functions, expressions, aggregates, Max
-from django.db.models.fields.json import KeyTextTransform
+from django.db.models import aggregates
 from django.http import HttpResponseRedirect
 from django.shortcuts import render
 from django.utils.decorators import method_decorator
 from django.utils.safestring import mark_safe
 from django.urls import reverse
-from django.views.generic.list import ListView
-from django.views.generic.detail import SingleObjectMixin
 from django.views.generic import DetailView
 from django.views.generic.edit import (
     CreateView,
     UpdateView,
-    View,
     FormMixin,
     ProcessFormView,
 )
 import django_tables2
+from django_tables2.views import SingleTableMixin
+from django_filters.views import FilterView
 import requests
 
 from api_app.models import (
@@ -35,8 +32,18 @@ from api_app.models import (
     PUBLISHED_CODE,
     AVAILABLE_STATUSES,
 )
-from data_models.models import Campaign, Instrument, Platform, PlatformType, PartnerOrg
-from . import tables, forms, mixins
+from data_models.models import (
+    Campaign,
+    CollectionPeriod,
+    Deployment,
+    Instrument,
+    Platform,
+    PlatformType,
+    PartnerOrg,
+    IOP,
+    SignificantEvent,
+)
+from . import tables, forms, mixins, filters
 
 
 @login_required
@@ -72,26 +79,21 @@ def deploy_admin(request):
 
 
 @method_decorator(login_required, name="dispatch")
-class ChangeSummaryView(django_tables2.SingleTableView):
+class SummaryView(django_tables2.SingleTableView):
     model = Change
-    model_names = [
-        M._meta.model_name for M in (Campaign, Platform, Instrument, PartnerOrg)
-    ]
+    models = (Campaign, Platform, Instrument, PartnerOrg)
     table_class = tables.ChangeSummaryTable
     paginate_by = 10
     template_name = "api_app/summary.html"
 
     def get_queryset(self):
         return (
-            Change.objects.filter(
-                content_type__model__in=self.model_names, action=CREATE
-            )
+            Change.objects.of_type(*self.models)
+            .filter(action=CREATE)
             # Prefetch related ContentType (used when displaying output model type)
             .select_related("content_type")
-            # Add last related ApprovalLog's date
-            .annotate(updated_at=aggregates.Max("approvallog__date")).order_by(
-                "-updated_at"
-            )
+            .add_updated_at()
+            .order_by("-updated_at")
         )
 
     def get_draft_status_count(self):
@@ -103,17 +105,17 @@ class ChangeSummaryView(django_tables2.SingleTableView):
 
         # Setup dict with 0 counts
         review_counts = {
-            model: {
+            model._meta.model_name: {
                 status.replace(" ", "_"): 0 for status in status_translations.values()
             }
-            for model in self.model_names
+            for model in self.models
         }
 
         # Populate with actual counts
         model_status_counts = (
-            Change.objects.filter(
+            Change.objects.of_type(*self.models)
+            .filter(
                 action=CREATE,
-                content_type__model__in=self.model_names,
                 status__in=status_ids,
             )
             .values_list("content_type__model", "status")
@@ -132,22 +134,19 @@ class ChangeSummaryView(django_tables2.SingleTableView):
             "draft_status_counts": self.get_draft_status_count(),
             "activity_list": ApprovalLog.objects.prefetch_related(
                 "change__content_type", "user"
-            ).order_by("-date")[:self.paginate_by / 2],
+            ).order_by("-date")[: self.paginate_by / 2],
         }
 
 
 @method_decorator(login_required, name="dispatch")
-class ChangeListView(django_tables2.SingleTableView):
+class CampaignListView(SingleTableMixin, FilterView):
     model = Change
-    table_class = tables.CampaignChangeListTable
     template_name = "api_app/change_list.html"
+    table_class = tables.CampaignChangeListTable
+    filterset_class = filters.ChangeStatusFilter
 
     def get_queryset(self):
-        return (
-            Change.objects.filter(content_type__model="campaign", action=CREATE)
-            # Add last related ApprovalLog's date
-            .annotate(updated_at=aggregates.Max("approvallog__date"))
-        )
+        return Change.objects.of_type(Campaign).filter(action=CREATE).add_updated_at()
 
     def get_context_data(self, **kwargs):
         return {
@@ -158,90 +157,77 @@ class ChangeListView(django_tables2.SingleTableView):
 
 
 @method_decorator(login_required, name="dispatch")
-class ChangeDetailView(DetailView):
+class CampaignDetailView(DetailView):
     model = Change
-    template_name = "api_app/change_detail.html"
-    queryset = Change.objects.filter(content_type__model=Campaign._meta.model_name)
+    template_name = "api_app/campaign_detail.html"
+    queryset = Change.objects.of_type(Campaign)
 
     def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
         deployments = (
-            Change.objects.filter(
-                content_type__model="deployment",
+            Change.objects.of_type(Deployment)
+            .filter(
                 update__campaign=str(self.kwargs[self.pk_url_kwarg]),
             )
-            .prefetch_related(
-                models.Prefetch(
-                    "approvallog_set",
-                    queryset=ApprovalLog.objects.order_by("-date").select_related(
-                        "user"
-                    ),
-                    to_attr="approvals",
-                )
-            )
+            .prefetch_approvals()
             .order_by(self.get_ordering())
         )
+        collection_periods = (
+            Change.objects.of_type(CollectionPeriod)
+            .filter(
+                update__deployment__in=[str(d.uuid) for d in deployments],
+            )
+            .select_related("content_type")
+            .prefetch_approvals()
+            .annotate_with_identifier_from_model(
+                model=Platform,
+                uuid_from="platform",
+                to_attr="platform_name",
+            )
+        )
+
+        # Build collection periods instruments (too difficult to do in SQL)
+        instrument_uuids = set(
+            uuid
+            for instruments in collection_periods.values_list(
+                "update__instruments", flat=True
+            )
+            for uuid in instruments
+        )
+        instrument_names = {
+            str(uuid): short_name
+            for uuid, short_name in Change.objects.of_type(Instrument)
+            .filter(uuid__in=instrument_uuids)
+            .values_list("uuid", "update__short_name")
+        }
+        for cp in collection_periods:
+            cp.instrument_names = sorted(
+                instrument_names.get(uuid) for uuid in cp.update.get("instruments")
+            )
+
         return {
-            **super().get_context_data(**kwargs),
+            **context,
             "deployments": deployments,
             "transition_form": forms.TransitionForm(
-                change=self.get_object(), user=self.request.user
+                change=context["object"], user=self.request.user
             ),
             "significant_events": (
-                Change.objects.select_related("content_type")
+                Change.objects.of_type(SignificantEvent)
                 .filter(
-                    content_type__model__iexact="significantevent",
                     update__deployment__in=[str(d.uuid) for d in deployments],
                 )
-                .prefetch_related(
-                    models.Prefetch(
-                        "approvallog_set",
-                        queryset=ApprovalLog.objects.order_by("-date").select_related(
-                            "user"
-                        ),
-                        to_attr="approvals",
-                    )
-                )
-                # Add related IOP's short_name if it exists
-                .annotate(
-                    iop_uuid=functions.Cast(
-                        KeyTextTransform("iop", "update"), models.UUIDField()
-                    ),
-                    iop_name=expressions.Subquery(
-                        Change.objects.filter(
-                            content_type__model__iexact="iop",
-                            uuid=expressions.OuterRef("iop_uuid"),
-                        ).values("update__short_name")[:1]
-                    ),
-                )
+                .select_related("content_type")
+                .prefetch_approvals()
             ),
-            "iops": Change.objects.select_related("content_type")
-            .filter(
-                content_type__model__iexact="iop",
-                update__deployment__in=[str(d.uuid) for d in deployments],
-            )
-            .prefetch_related(
-                models.Prefetch(
-                    "approvallog_set",
-                    queryset=ApprovalLog.objects.order_by("-date").select_related(
-                        "user"
-                    ),
-                    to_attr="approvals",
+            "iops": (
+                Change.objects.of_type(IOP)
+                .filter(
+                    update__deployment__in=[str(d.uuid) for d in deployments],
                 )
+                .select_related("content_type")
+                .prefetch_approvals()
             ),
-            "collection_periods": Change.objects.select_related("content_type")
-            .filter(
-                content_type__model__iexact="collectionperiod",
-                update__deployment__in=[str(d.uuid) for d in deployments],
-            )
-            .prefetch_related(
-                models.Prefetch(
-                    "approvallog_set",
-                    queryset=ApprovalLog.objects.order_by("-date").select_related(
-                        "user"
-                    ),
-                    to_attr="approvals",
-                )
-            ),
+            "collection_periods": collection_periods,
         }
 
     def get_ordering(self):
@@ -250,7 +236,6 @@ class ChangeDetailView(DetailView):
 
 @method_decorator(login_required, name="dispatch")
 class ChangeCreateView(mixins.ChangeModelFormMixin, CreateView):
-
     model = Change
     fields = ["content_type", "model_instance_uuid", "action", "update"]
     template_name = "api_app/change_add_form.html"
@@ -272,7 +257,7 @@ class ChangeCreateView(mixins.ChangeModelFormMixin, CreateView):
         }
 
     def get_success_url(self):
-        return reverse("change-form", args=[self.object.pk])
+        return reverse("mi-change-update", args=[self.object.pk])
 
     def get_model_form_content_type(self) -> ContentType:
         if not hasattr(self, "model_form_content_type"):
@@ -297,15 +282,16 @@ class ChangeCreateView(mixins.ChangeModelFormMixin, CreateView):
 
 @method_decorator(login_required, name="dispatch")
 class ChangeUpdateView(mixins.ChangeModelFormMixin, UpdateView):
-    success_url = "/"
     fields = ["content_type", "model_instance_uuid", "action", "update", "status"]
-
     prefix = "change"
     template_name = "api_app/change_form.html"
 
+    def get_success_url(self):
+        return reverse("mi-change-update", args=[self.object.pk])
+
     def get_queryset(self):
         # Prefetch content type for performance
-        return Change.objects.select_related("content_type")
+        return Change.objects.select_related("content_type").prefetch_approvals()
 
     def get_context_data(self, **kwargs):
         return {
@@ -313,10 +299,6 @@ class ChangeUpdateView(mixins.ChangeModelFormMixin, UpdateView):
             "transition_form": forms.TransitionForm(
                 change=self.get_object(), user=self.request.user
             ),
-            # Add approvals to context
-            "approvals": self.get_object()
-            .approvallog_set.order_by("-date")
-            .prefetch_related("user"),
         }
 
     def get_model_form_content_type(self) -> ContentType:
@@ -335,26 +317,21 @@ class ChangeUpdateView(mixins.ChangeModelFormMixin, UpdateView):
 
 
 @method_decorator(login_required, name="dispatch")
-class PlatformListView(django_tables2.SingleTableView):
+class PlatformListView(SingleTableMixin, FilterView):
     model = Change
-    table_class = tables.PlatformChangeListTable
     template_name = "api_app/change_list.html"
+    table_class = tables.PlatformChangeListTable
+    filterset_class = filters.ChangeStatusFilter
 
     def get_queryset(self):
         return (
-            Change.objects.filter(content_type__model="platform", action=CREATE)
-            # Add last related ApprovalLog's date
-            .annotate(updated_at=aggregates.Max("approvallog__date"))
-            # Add related PlatformType's short_name
-            .annotate(
-                platform_type_uuid=functions.Cast(
-                    KeyTextTransform("platform_type", "update"), models.UUIDField()
-                ),
-                platform_type_name=expressions.Subquery(
-                    PlatformType.objects.filter(
-                        uuid=expressions.OuterRef("platform_type_uuid")
-                    ).values("short_name")[:1]
-                ),
+            Change.objects.of_type(Platform)
+            .filter(action=CREATE)
+            .add_updated_at()
+            .annotate_with_identifier_from_model(
+                model=PlatformType,
+                uuid_from="platform_type",
+                to_attr="platform_type_name",
             )
         )
 
@@ -362,39 +339,37 @@ class PlatformListView(django_tables2.SingleTableView):
         return {
             **super().get_context_data(**kwargs),
             "display_name": "Platform",
-            "model": "platform",
+            "model": Platform._meta.model_name,
         }
 
 
 @method_decorator(login_required, name="dispatch")
-class InstrumentListView(django_tables2.SingleTableView):
+class InstrumentListView(SingleTableMixin, FilterView):
     model = Change
-    table_class = tables.BasicChangeListTable
     template_name = "api_app/change_list.html"
+    table_class = tables.BasicChangeListTable
+    filterset_class = filters.ChangeStatusFilter
 
     def get_queryset(self):
-        return Change.objects.filter(
-            content_type__model="instrument", action=CREATE
-        ).annotate(updated_at=Max("approvallog__date"))
+        return Change.objects.of_type(Instrument).filter(action=CREATE).add_updated_at()
 
     def get_context_data(self, **kwargs):
         return {
             **super().get_context_data(**kwargs),
             "display_name": "Instrument",
-            "model": "instrument",
+            "model": Instrument._meta.model_name,
         }
 
 
 @method_decorator(login_required, name="dispatch")
-class PartnerOrgListView(django_tables2.SingleTableView):
+class PartnerOrgListView(SingleTableMixin, FilterView):
     model = Change
-    table_class = tables.BasicChangeListTable
     template_name = "api_app/change_list.html"
+    table_class = tables.BasicChangeListTable
+    filterset_class = filters.ChangeStatusFilter
 
     def get_queryset(self):
-        return Change.objects.filter(
-            content_type__model=PartnerOrg._meta.model_name, action=CREATE
-        ).annotate(updated_at=Max("approvallog__date"))
+        return Change.objects.of_type(PartnerOrg).filter(action=CREATE).add_updated_at()
 
     def get_context_data(self, **kwargs):
         return {
@@ -402,11 +377,6 @@ class PartnerOrgListView(django_tables2.SingleTableView):
             "display_name": "Partner Organization",
             "model": PartnerOrg._meta.model_name,
         }
-
-
-@login_required
-def to_be_developed(request):
-    return render(request, "api_app/to_be_developed.html")
 
 
 @method_decorator(login_required, name="dispatch")