+++ cmr/doi_matching.py
@@ -1,23 +1,33 @@
-from datetime import datetime
+import json
 import pickle
+from datetime import datetime
 
-from api_app.models import Change
+from api_app.models import (AWAITING_ADMIN_REVIEW_CODE, AWAITING_REVIEW_CODE,
+                            CREATE, CREATED_CODE, DELETE, IN_ADMIN_REVIEW_CODE,
+                            IN_PROGRESS_CODE, IN_REVIEW_CODE, PUBLISHED_CODE,
+                            UPDATE, Change)
 from data_models.models import DOI
+from django.apps import apps
+from django.contrib.contenttypes.models import ContentType
+from django.core import serializers
 
-from cmr.api import Api
 from cmr.cmr import query_and_process_cmr
-from cmr.utils import (
-    clean_table_name,
-    purify_list
-)
+from cmr.utils import clean_table_name, purify_list
 
+ALL_STATUSES = [
+    CREATED_CODE,
+    IN_PROGRESS_CODE,
+    AWAITING_REVIEW_CODE,
+    IN_REVIEW_CODE,
+    AWAITING_ADMIN_REVIEW_CODE,
+    IN_ADMIN_REVIEW_CODE,
+    PUBLISHED_CODE
+]
 
 class DoiMatcher():
     def __init__(self):
         self.uuid_to_aliases = {}
         self.table_to_valid_uuids = {}
-        self.change_requests = []
-        self.api = Api()
 
 
     def universal_get(self, table_name, uuid):
@@ -26,84 +36,28 @@ class DoiMatcher():
         results from the main db.
 
         Args:
-            api (class): api class
             table_name (str): table name such as 'gcmd_platform'
             uuid (str): uuid of the object
 
         Returns:
-            data (dict): object from change table or None if not found
+            data (dict): object from db or change table if not found
         """
 
-        db_response = self.api.get(f'{table_name}/{uuid}')
-        if db_response['success']:
-            data = db_response['data']
-        else:
-            draft_response = self.api.get(f'change_request/{uuid}')
-            if draft_response['success']:
-                data = draft_response['data']['update']
-                data['uuid'] = draft_response['data']['uuid']
-                data['change_object'] = True
-                data['change_object_action'] = draft_response['data']['action']
-                data['change_object_status'] = draft_response['data']['status']
-            else:
-                data = None
+        model = apps.get_model('data_models', table_name.replace('_', ''))
+        # attempt to find the uuid as a published object
+        try:
+            obj = model.objects.get(uuid=uuid)
+            data = json.loads(serializers.serialize('json', [obj,]))[0]['fields']
+
+        # if the published object isn't found, search the drafts
+        except model.DoesNotExist:
+            model = apps.get_model('api_app', 'change')
+            obj = model.objects.get(uuid=uuid)
+            data = json.loads(serializers.serialize('json', [obj,]))[0]['fields']['update']
+            data['uuid'] = uuid
+            data['change_object'] = True
+
         return data
-
-
-    @staticmethod
-    def filter_change_object(co, action=None, statuses=None, table_name=None, query_parameter=None, query_value=None):
-        """Inputs a single change object and returns a boolean indicating whether it passed
-        the filter parameters. This is meant to be run on each change object in order to gather
-        all the change objects that match a certain criteria, such as all create objects from
-        the season table with a short_name of fall.
-
-        Args:
-            co (dict): Change object dictionary
-            action (str, optional): string from 'create', 'delete', 'update'. This indicates
-            the type of change object. Defaults to None.
-            statuses (list[int], optional): List of integers indicating the status of the
-                change object in the approval process. Defaults to None.
-            table_name (str, optional): Table name such as season. Defaults to None.
-            query_parameter (str, optional): Name of the parameter to search. This should be
-                a field in the object such as short_name or uuid. Defaults to None.
-            query_value (str, optional): This is the expected value for the query_parameter,
-                such as fall for short_name. Defaults to None.
-
-        Returns:
-            bool: Boolean indicating whether the change object matched the filter values.
-        """
-
-        # tests will default to True if the associated arg is not passed
-        is_action, is_status, is_table_name, is_value = True, True, True, True
-
-        if action:
-            is_action = co['action'].lower()==action.lower()
-
-        if statuses:
-            is_status = co['status'] in statuses
-
-        if table_name:
-            is_table_name = clean_table_name(co['model_name']) == clean_table_name(table_name)
-
-        if query_value:
-            is_value = co['update'].get(query_parameter, '').lower() == query_value.lower()
-
-        return is_action and is_status and is_table_name and is_value
-
-
-    def _get_change_requests(self):
-        """Downloads all change requests from the database. This query is only executed once
-        during the lifetime of a DoiMatcher object, under the assumption that there will be
-        no meaningful changes to the db during runtime, and because this greatly speeds up
-        the overall process.
-
-        Returns:
-            change_requests (list): List of all change_requests from db
-        """
-
-        if not self.change_requests:
-            self.change_requests = self.api.get('change_request')['data']
-        return self.change_requests
 
 
     def valid_object_list_generator(self, table_name, query_parameter=None, query_value=None):
@@ -123,19 +77,21 @@ class DoiMatcher():
             uuid_list (list): List of strings of uuids for the valid objects from a table
         """
 
-        # this improves speed during development
-        change_requests = self._get_change_requests()
+        valid_objects = Change.objects.filter(
+            content_type__model=table_name,
+            action=CREATE
+            ).exclude(
+                action=DELETE,
+                status=PUBLISHED_CODE)
 
-        # get all create items of any approval status from the given table
-        created = [c for c in change_requests if self.filter_change_object(c, 'create', [1,2,3], table_name, query_parameter, query_value)]
-        # get all approved deleted objects
-        deleted = [c for c in change_requests if self.filter_change_object(c, 'delete', [3], table_name, query_parameter, query_value)]
+        if query_parameter:
+            query_parameter = 'update__' + query_parameter
+            kwargs = {query_parameter: query_value}
+            valid_objects = valid_objects.filter(**kwargs)
 
-        # filter out the approved deletes from the approved and in progress create list
-        deleted_uuids = [d['model_instance_uuid'] for d in deleted]
-        valid_objects = [c for c in created if c['uuid'] not in deleted_uuids]
+        valid_object_uuids = [str(uuid) for uuid in valid_objects.values_list('uuid', flat=True)]
 
-        return [o['uuid'] for o in valid_objects]
+        return valid_object_uuids
 
 
     def universal_alias(self, table_name, uuid):
@@ -348,12 +304,21 @@ class DoiMatcher():
 
         # search db for existing items with concept_id
         existing_doi_uuids = self.valid_object_list_generator('doi', query_parameter='concept_id', query_value=doi['concept_id'])
-        if len(existing_doi_uuids)>1:
-            raise ValueError('There has been an internal database error')
+        # this check can fail for some complicated reasons that will be addressed in a future PR
+        # if len(existing_doi_uuids)>1:
+        #     raise ValueError('There has been an internal database error')
 
         # if none exist add normally as a draft
         if not existing_doi_uuids:
-            self.api.create('doi', doi, draft=True)
+            doi_obj = Change(
+                content_type=ContentType.objects.get(model='doi'),
+                model_instance_uuid=None,
+                update=json.loads(json.dumps(doi)),
+                status=CREATED_CODE,
+                action=CREATE
+            )
+            doi_obj.save()
+
             return 'Draft created for DOI'
 
         uuid = existing_doi_uuids[0]
@@ -364,7 +329,7 @@ class DoiMatcher():
                 doi[field].extend(existing_doi.get(field))
                 doi[field] = list(set(doi[field]))
 
-            draft = Change.objects.all().filter(uuid=uuid).first()
+            draft = Change.objects.get(uuid=uuid)
             draft.update = doi
             draft.save()
 
@@ -385,7 +350,16 @@ class DoiMatcher():
         for field in ['campaigns', 'instruments', 'platforms', 'collection_periods']:
             doi[field] = list(set(doi[field]))
 
-        self.api.update(f'doi/{uuid}', data=doi, draft=True)
+        doi_obj = Change(
+            content_type=ContentType.objects.get(model='doi'),
+            model_instance_uuid=str(uuid),
+            update=json.loads(json.dumps(doi)),
+            status=CREATED_CODE,
+            action=UPDATE
+        )
+
+        doi_obj.save()
+
         return f'DOI already exists in database. Update draft created. {uuid}'
 
 
@@ -421,6 +395,8 @@ class DoiMatcher():
 
         if failed or not development:
             metadata_list = query_and_process_cmr(table_name, aliases)
+        
+        if development:
             pickle.dump(metadata_list, open(f'metadata_{uuid}', 'wb'))
 
         supplemented_metadata_list = self.supplement_metadata(metadata_list, development)