+++ config/settings/local.py
@@ -67,3 +67,33 @@ if os.name == 'nt':
     os.environ['GDAL_DATA'] = OSGEO4W + r"\share\gdal"
     os.environ['PROJ_LIB'] = OSGEO4W + r"\share\proj"
     os.environ['PATH'] = OSGEO4W + r"\bin;" + os.environ['PATH']
+
+# everything below this line is required to get shell_plus --notebook to work from within a docker container
+SHELL_PLUS = "ipython"
+
+SHELL_PLUS_PRINT_SQL = True
+
+NOTEBOOK_ARGUMENTS = [
+    "--ip",
+    "0.0.0.0",
+    "--port",
+    "8888",
+    "--allow-root",
+    "--no-browser",
+]
+
+IPYTHON_ARGUMENTS = [
+    "--ext",
+    "django_extensions.management.notebook_extension",
+    "--debug",
+]
+
+IPYTHON_KERNEL_DISPLAY_NAME = "Django Shell-Plus"
+
+SHELL_PLUS_POST_IMPORTS = [ # extra things to import in notebook
+    ("module1.submodule", ("func1", "func2", "class1", "etc")),
+    ("module2.submodule", ("func1", "func2", "class1", "etc"))
+
+]
+
+os.environ["DJANGO_ALLOW_ASYNC_UNSAFE"] = "true" # only use in development